+++ src/main/java/it/pagopa/ecommerce/payment/methods/infrastructure/PaymentMethodDocument.java
@@ -7,7 +7,6 @@ import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.util.Pair;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
@@ -22,7 +21,7 @@ public class PaymentMethodDocument {
     private String paymentMethodDescription;
     private String paymentMethodStatus;
     private String paymentMethodAsset;
-    private List<Pair<BigInteger, BigInteger>> paymentMethodRanges;
+    private List<Pair<Long, Long>> paymentMethodRanges;
     @Indexed(unique = true)
     private String paymentMethodTypeCode;
 }
